Add server, GET api feature and index.html Refactor crawling codes into crawler folder
